Split TotalDurationChart
